requirements.txt
Django==3.0.7
djangorestframework==3.10.3
django-rest-auth==0.9.5
django-allauth==0.40.0
psycopg2==2.8.4
djangorestframework-jwt==1.11.0
dj-static==0.0.6